requirements.txt
psycopg[binary]==3.2.1
python-dotenv==1.0.1
requests==2.32.3
